Add exportMosaic script
With this script it is possible to export a mosaic in the format needed
for the website. As results of running the script you get two 'files',
one .js file which is used directly for visualizing the mosaic on the website
and another .kml file which you can provide to the user for download.
Requires IITC with installed mission plugin.