Update map and overview to show mission pics directly
With this update it is possible to use the mission pictures almost directly
on the website. There is no need to create an whole mission picture e.g.
from the scanner anymore.
With this change the lazyload.js is included for a hughly improved
loading performance on the overview page, because otherwise
loading a couple hundred pictures takes a really long time and
leads to bad user experience.